+++ src/Data/Morpheus/Types.hs
@@ -35,6 +35,7 @@ module Data.Morpheus.Types
   , publish
   , subscribe
   , unsafeInternalContext
+  , SubField
   )
 where
 
@@ -69,6 +70,7 @@ import           Data.Morpheus.Types.Internal.Resolving
                                                 , PushEvents(..)
                                                 , subscribe
                                                 , unsafeInternalContext
+                                                , UnSubResolver
                                                 )
 import           Data.Morpheus.Types.IO         ( GQLRequest(..)
                                                 , GQLResponse(..)
@@ -88,6 +90,8 @@ type ResolveQ e m a = Res e m (a (Res e m))
 type ResolveM e m a = MutRes e m (a (MutRes e m))
 type ResolveS e m a = SubRes e m (a (Res e m))
 
+-- Subsciption Object Resolver Fields
+type SubField m a = (m (a (UnSubResolver m))) 
 
 publish :: Monad m => [e] -> Resolver MUTATION e m ()
 publish = pushEvents