+++ changelog.md
@@ -58,7 +58,7 @@
   }
   ```
 
-- `unsafeInternalContext` to get resolver context, use only if it real necessary.
+- `unsafeInternalContext` to get resolver context, use only if it really necessary.
   the code depend on it may break even on minor version changing, therefore is called `unsafe`.
   
   ```hs